Add delete relation in databse and quick code refactor
