+++ WTW_app/details/details_repository.py
